mex/source/simulate: changed some cast to (long int) for compatibility with x64
git-svn-id: https://www.dynare.org/svn/dynare/dynare_v4@1680 ac1d8469-bf42-47a9-8791-bf33cf982152

+++ matlab/mex/sources/simulate/Interpreter.cc
@@ -1141,7 +1141,7 @@ Interpreter::compute_blocks(string file_name, string bin_basename)
             mexPrintf("FBEGINBLOCK\n");
             mexEvalString("drawnow;");
 #endif
-            lBlock.begin=get_code_pos-int(Init_Code);
+            lBlock.begin=get_code_pos-(long int)(Init_Code);
 #ifdef DEBUGC
             mexPrintf("Block[Block_Count].begin=%d\n",lBlock.begin);
             mexEvalString("drawnow;");
@@ -1194,7 +1194,7 @@ Interpreter::compute_blocks(string file_name, string bin_basename)
             T=(double*)mxMalloc(var*(periods+y_kmin+y_kmax)*sizeof(double));
             break;
           default :
-            mexPrintf("Unknow command : %d at pos %d !!\n",int(code),int(get_code_pos)-int(Init_Code));
+            mexPrintf("Unknow command : %d at pos %d !!\n",(long int)(code),(long int)(get_code_pos)-(long int)(Init_Code));
             mexEvalString("st=fclose('all');clear all;");
             mexEvalString("drawnow;");
             mexErrMsgTxt("End of simulate");
