+++ matlab/mex/sources/simulate/Interpreter.hh
@@ -43,7 +43,7 @@ typedef struct Block_contain_type
 
 typedef struct Block_type
 {
-  int begin, end, size, type;
+  long int begin, end, size, type;
 };
 
 #define pow_ pow
@@ -52,7 +52,7 @@ typedef struct Block_type
 #define get_code_pdouble      (double*)Code; Code+=sizeof(double);
 #define get_code_bool         *((bool*)(Code++))
 #define get_code_char         *((char*)(Code++))
-#define get_code_pos          (int)Code
+#define get_code_pos          (long int)Code
 #define get_code_pointer      Code
 #define set_code_pointer(pos) Code=pos
